feat(catalogue): add support for linking catalogues to a catalogue
This commit adds support for linking catalogues to a catalogue via the API. The `CatalogueLinkCataloguesCommand` command is used to link the specified catalogues to the target catalogue. The linked catalogues are specified as a list of catalogue identifiers. The `CatalogueLinkedCataloguesEvent` event is emitted after the linking operation is completed.

The scenario in the feature file demonstrates the usage of the API to link catalogues to a catalogue.