refactor(config): rename 'orchestrator' to 'admin' for clarity
The 'orchestrator' property is renamed to 'admin' across multiple files to better reflect its purpose and improve code readability. This change ensures that the naming is consistent with the role it plays in the application, which is related to administrative functions rather than orchestration.

fix(initScript): disable cccev and catalogue by default

The default values for 'cccev' and 'catalogue' are changed to 'false' in the InitScript. This change is made to prevent unnecessary initialization of these components when they are not needed, optimizing the script's performance.

feat(catalogue): add catalogueStandards function

A new function, 'catalogueStandards', is added to define a catalogue of standards. This function provides a structured way to explore recognized standards for environmental project evaluation, enhancing the application's capability to manage and present standard-related data.

fix(TransactionPdfCertificate): set fixed parentWidth for PdfDisplayer

The 'parentWidth' for the PdfDisplayer component is set to a fixed value of 800. This change ensures consistent rendering of PDF documents, regardless of the element size, improving the user experience by maintaining a stable layout.