feat(docker-compose): add connect-admin service and configuration
Introduce the connect-admin service to the development environment,
enabling management of administrative tasks. Update the .env_dev file
to include the connect-admin service in the DOCKER_COMPOSE_FILE and
add relevant environment variables for configuration. Create new
configuration files for OIDC trusted domains and web defaults to
support the connect-admin service. Update the space-config.json to
include the connect-admin client ID and permissions. Modify the
dev-compose.mk to process the new configuration files. Add a new
docker-compose-connect-admin.yml file to define the connect-admin
service.

These changes are made to support the addition of the connect-admin
service, which is necessary for handling administrative functions
within the platform. The new configurations and permissions ensure
proper integration and security management for the service.