feat(api): update query keys to include project and data prefixes for better organization fix(OidcTrustedDomains.js): add new trusted OIDC domain for improved authentication chore(env-config.js): remove obsolete env-config.js file as its functionality is no longer needed
The changes to the query keys enhance the organization and clarity of the API endpoints by categorizing them under relevant prefixes. Adding a new trusted OIDC domain improves the security and functionality of the authentication process. The removal of the env-config.js file eliminates unnecessary code, streamlining the project structure.