feat(env): add KC_REALM_DISPLAY_NAME to environment configuration for better clarity fix(space-config.json): rename "space" to "identifier" for consistency in configuration fix(space-create.json): add displayName to space creation configuration for better identification refactor(permissions): update permissions structure to include new roles and improve clarity fix(ActorBuilder.kt): change role identifier to string for consistency feat(LoginConfigTotp.tsx): add userLabel for better user experience in TOTP configuration style(CatalogueViewPage.tsx): import DatasetDataSection for improved functionality
New environment variable KC_REALM_DISPLAY_NAME provides a clearer name for the realm. The configuration files are updated to ensure consistency and clarity in naming conventions. The permissions structure is refined to include new roles, enhancing the overall organization of permissions. Changes in the ActorBuilder and LoginConfig components improve consistency and user experience.